handler modificado de acordo com epos2, porém no epos1 havia uma definição de função

+++ include/ic.h
@@ -13,7 +13,7 @@ protected:
     IC_Common() {}
 
 public:
-    typedef unsigned int Interrupt_Id;
+    typedef long unsigned int Interrupt_Id;
     typedef void (* Interrupt_Handler)(const Interrupt_Id &);
 };
